Moved the find squares loop
I just made a small edit that moves the find squares loop outside of the image processing loop. I was using this function and noticed that it outputted a lot of copies of the same image, because of the location of the loops. It worked just as well and returned no duplicates.